amended some of the code in virtual voice
- virtual voice's threadplayhead variable was still a type size_t instead of a float

+++ Maginengine Audio/VirtualVoice.cpp
@@ -29,7 +29,7 @@ void VirtualVoice::processAudio(float* outputBuffer, ma_uint32 frameCount)
 		
 		for (ma_uint32 i = 0; i < frameCount; ++i)
 		{
-			size_t threadPlayhead = playHead.load();
+			float threadPlayhead = playHead.load();
 			if (threadPlayhead < buffer.size())
 			{
 				threadPlayhead++;
